requirements.txt
google-generativeai==0.3.2
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.14
requests==2.31.0
python-dotenv==1.0.0
pywhatkit==5.4
pyautogui==0.9.54
wikipedia-api==0.6.0
